requirements.txt
tensorflow
streamlit
pandas
numpy
pillow